Updated example to show error handling.
I mainly did this to test a thought I had about scoping issues we would
have passing in the OnErrorMethod.

When a variable is assigned a pointer to a function on an object, and
then called, it executes in a different context.

So, if your error handler is not completely self contained then, you'll
probably want to create a simple proxy method.