Added payload arg name config and scope merge config
Added ability to specify the argument name for the HTTP request body
value that is passed to your service method. (Still defaults to
"Payload".)
Also, added ability to control which scopes are merged into the
arguments scope that is passed to your service method. (Still defaults
to all of them.)